fix(web): fix mobile layout issues and add i18n for pricing section
- Fix mobile navbar overflow by hiding login link on small screens
- Fix pricing billing toggle overlapping hosting tabs on mobile (stack vertically)
- Rewrite footer with separate mobile/desktop layouts for better UX
- Change French nav CTA from "Essayer gratuitement" to "Démarrer"
- Replace all hardcoded English strings in PricingSection with tPricing() calls
- Add missing i18n keys to en/fr/es/zh pricing locale files
- Fix server/workspace/member feature keys to use raw string keys (upTo1, etc.)
