+++ apps/web/public/locales/es/pricing.json
@@ -16,6 +16,11 @@
   "coreFeatures": "Funcionalidades principales",
   "support": "Soporte",
   "enterpriseFeatures": "Funcionalidades empresariales",
+  "mostPopular": "Más popular",
+  "freeTrial": "Prueba gratuita",
+  "contactSales": "Contactar ventas",
+  "soon": "Próximamente",
+  "securityCompatibility": "Seguridad & compatibilidad",
   "plans": {
     "starter": {
       "name": "Starter",
@@ -74,7 +79,24 @@
     "customLicenseTiers": "Licencias personalizadas",
     "customLicenseTiersDesc": "Opciones de licencia flexibles",
     "completeDataControl": "Control total de datos",
-    "completeDataControlDesc": "Todos los datos permanecen en tu infraestructura"
+    "completeDataControlDesc": "Todos los datos permanecen en tu infraestructura",
+    "queueExchangeManagement": "Gestión de colas, exchanges, VHosts y usuarios",
+    "alertsWebhooks": "Alertas & webhooks",
+    "topologyVisualization": "Visualización de topología",
+    "roleBasedAccess": "Acceso basado en roles",
+    "ltsVersions": "Versiones LTS de RabbitMQ",
+    "allLtsVersions": "Todas las versiones LTS 3.x y 4.x",
+    "allRabbitMQVersions": "Todas las versiones de RabbitMQ",
+    "ssoSamlOidc": "SSO, SAML & OIDC",
+    "server_upTo1": "1 servidor RabbitMQ",
+    "server_upTo3": "Hasta 3 servidores RabbitMQ",
+    "server_unlimited": "Servidores RabbitMQ ilimitados",
+    "workspace_upTo1": "1 espacio de trabajo",
+    "workspace_upTo3": "Hasta 3 espacios de trabajo",
+    "workspace_unlimited": "Espacios de trabajo ilimitados",
+    "member_upTo1": "1 miembro de equipo",
+    "member_upTo3": "Hasta 3 miembros de equipo",
+    "member_unlimited": "Miembros de equipo ilimitados"
   },
   "limits": {
     "upTo1": "Hasta 1",
