+++ apps/web/public/locales/fr/nav.json
@@ -3,6 +3,6 @@
   "features": "Fonctionnalités",
   "pricing": "Tarifs",
   "login": "Connexion",
-  "tryForFree": "Essayer gratuitement",
+  "tryForFree": "Démarrer",
   "whatsNew": "Nouveautés"
 }
